Removes colon from library table cell if no subtitle follows (#633).
